Flytter oppdatering av andeler på TilkjentYtelse ut av EØS og nærmere kjernen
